Added enemy pathfinding towards the player. Fixed the enemy movement to allow movement around obstacles
